Services/html_deploy/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-storage==2.16.0
google-auth==2.29.0
python-multipart==0.0.9
itsdangerous==2.2.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
